requirements.txt
altair==5.5.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.2.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
contourpy==1.3.2
curl_cffi==0.12.0
cycler==0.12.1
dash==3.2.0
debugpy==1.8.15
decorator==5.2.1
delta-spark==3.0.0
dnspython==1.16.0
docopt==0.6.2
dotenv==0.9.9
emoji==2.14.1
exceptiongroup==1.3.0
executing==2.2.0
Flask==3.1.2
fonttools==4.59.1
frozendict==2.4.6
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hdfs==2.7.3
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.0
ipython==8.37.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kafka-python==2.2.15
kiwisolver==1.4.9
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
multitasking==0.0.12
narwhals==2.2.0
neo4j==5.28.2
nest-asyncio==1.6.0
numpy==1.26.4
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
parso==0.8.4
peewee==3.18.2
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
plotly==6.3.0
praw==7.8.1
prawcore==2.4.0
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==21.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.2
pymongo==3.11.0
pyparsing==3.2.3
PySocks==1.7.1
pyspark==3.5.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyzmq==27.0.0
RapidFuzz==3.13.0
redis==6.2.0
referencing==0.36.2
requests==2.32.4
retrying==1.4.2
rpds-py==0.27.0
seaborn==0.13.2
selenium==4.34.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
stack-data==0.6.3
streamlit==1.48.1
streamlit-autorefresh==1.0.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
update-checker==0.18.0
urllib3==2.5.0
vaderSentiment==3.3.2
watchdog==6.0.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
yfinance==0.2.65
zipp==3.23.0
